Improve filtering of large profiles
And some minor related cleanup to other UI filtering.

Resolves #20

+++ core/src/main/java/org/glowroot/collector/TraceBase.java
@@ -52,7 +52,7 @@ public abstract class TraceBase {
     public abstract @Nullable Long threadAllocatedBytes();
     @JsonRawValue
     public abstract @Nullable String gcInfos();
-    public abstract long entryCount();
+    public abstract int entryCount();
     public abstract long profileSampleCount();
     abstract Existence entriesExistence();
     abstract Existence profileExistence();
